Euler crap

+++ src/me/lukasmetlicka/compsci/plainclass/Shapes/Quadrilateral.java
@@ -26,7 +26,7 @@ public class Quadrilateral implements IShape {
 	}
 	@Override
 	public int getLength(){
-		return Math.sqrt((p1.getX()-p2.getX())*(p1.getX()-p2.getX())+(p1.getY()-p2.getY())*(p1.getY()-p2.getY()));
+		return (int) Math.sqrt((p1.getX() - p2.getX()) * (p1.getX() - p2.getX()) + (p1.getY() - p2.getY()) * (p1.getY() - p2.getY()));
 	}
 	public int getHeight(){
 		return Math.max(p1.getX(), p4.getX())- Math.min(p1.getX(), p4.getX());
